pyproject.toml
[tool.poetry]
name = "moex-fetch"
version = "0.1.0"
description = "Приложение для получения данных с Московской биржи"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
aiohttp = "^3.9.1"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
alembic = "^1.12.1"

[tool.poetry.scripts]
start = "run:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 